+++ src/_packages/node.index.ts
@@ -1,19 +1,19 @@
 // ⚠️ WARNING: This code has been generated so that any manual changes will be overwritten
 // `@promptbook/node`
 
-import { PROMPTBOOK_VERSION } from '../version';
 import { createCollectionFromDirectory } from '../collection/constructors/createCollectionFromDirectory';
 import { createLlmToolsFromConfigurationFromEnv } from '../llm-providers/_common/createLlmToolsFromConfigurationFromEnv';
 import { createLlmToolsFromEnv } from '../llm-providers/_common/createLlmToolsFromEnv';
-import { FilesStorage } from '../storage/files-storage/FilesStorage';
-
+import { FileCacheStorage } from '../storage/file-cache-storage/FileCacheStorage';
+import { PROMPTBOOK_VERSION } from '../version';
 
 // Note: Exporting version from each package
 export { PROMPTBOOK_VERSION };
 
-
 // Note: Entities of the `@promptbook/node`
-export { createCollectionFromDirectory };
-export { createLlmToolsFromConfigurationFromEnv };
-export { createLlmToolsFromEnv };
-export { FilesStorage };
+export {
+    createCollectionFromDirectory,
+    createLlmToolsFromConfigurationFromEnv,
+    createLlmToolsFromEnv,
+    FileCacheStorage,
+};